+++ src/Aonik.Domain/Autonumbering/Entities/AutonumberResetPolicy.cs
@@ -1,0 +1,8 @@
+namespace Aonik.Domain.Autonumbering.Entities;
+
+public enum AutonumberResetPolicy
+{
+    None = 0,
+    Monthly = 1,
+    Yearly = 2
+}